feat(stream): infer the streamed delta element type from `output` (#115)
The streaming surfaces hardwired their delta element type — `sse` to
`SseEvent[]` (with `data: unknown`) and `stream` to `unknown[]` — so setting
`output` validated each delta at runtime but never refined the type the caller
sees. Make `output` refine the static element type.

This is PURELY a compile-time change: no runtime code path, decoded value, or
emitted event is touched. The built `lib/sse.mjs`/`lib/stream.mjs` are
byte-identical to main modulo minifier identifier renames, and the root-entry
bundle-size gate is unchanged.

- `SseEvent` is now generic (`SseEvent<T = unknown>`); the `= unknown` default
  keeps the parser, `contractValue`, and every existing import compiling
  unchanged. `sse(config)` returns `Stitch<SseEvent<OutputOf<C>>[], InputOf<C>>`
  — with no `output` that is `SseEvent<unknown>[]` (identical to today); with
  `output: S` each event's `.data` is the inferred type (the `sse` surface's
  `contractValue` validates `.data`).
- `stream`'s element type is decoder-dependent via a new local `StreamElement<C>`
  helper: `'lines'` → `string`, `'ndjson'` → `OutputOf<C>`, `'bytes'` (default)
  and any unknown decode → `Uint8Array`. `output` refines ONLY the structured
  `'ndjson'` decoder; `stream({ output: S })` with no `decode` stays
  `Uint8Array[]` (intended — documented in a comment).
- Threads `OutputOf<C>` through the `SseSeamApi`/`StreamSeamApi` interfaces, the
  standalone `sseStitch`/`streamStitch`, and both `bindSeam` inner `stitch`
  functions, keeping the existing `as unknown as Stitch<…>` retype pattern.
- Adds `test-d/streaming-inference.test-d.ts` (the proof; the change is invisible
  at runtime) and drops a now-unnecessary `as Uint8Array` cast in stream.spec.ts.

+++ packages/core/test/stream.spec.ts
@@ -45,7 +45,7 @@ describe('stream decoders (Decision 5)', () => {
 
         const chunks = await s();
         expect(chunks.every((c) => c instanceof Uint8Array)).toBe(true);
-        const joined = chunks.map((c) => td.decode(c as Uint8Array)).join('');
+        const joined = chunks.map((c) => td.decode(c)).join('');
         expect(joined).toBe('abcd');
     });
